# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocalMind API",
        "description": "This is an API documentation for LocalMind.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3535/api",
            "description": "LocalMind API Server"
        }
    ],
    "paths": {
...
        "/answers/{question}": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Store a new answer",
                "operationId": "109afcf91209d8a347dff8cb03a64a94",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "The ID of the question being answered",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answer": {
                                        "description": "The text content of the answer.",
                                        "type": "string",
                                        "example": "This is my answer content."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the answer was created successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Answer"
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Answer created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/answers/{answer}": {
            "put": {
                "tags": [
                    "Answers"
                ],
                "summary": "Update an existing answer",
                "operationId": "1b1e3a10e9b3b7abb726ce85ac63165c",
                "parameters": [
                    {
                        "name": "answer",
                        "in": "path",
                        "description": "The ID of the answer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answer": {
                                        "description": "The updated text content of the answer.",
                                        "type": "string",
                                        "example": "Updated answer text."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Answer"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Answer updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Answer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Answers"
                ],
                "summary": "Delete an answer",
                "operationId": "3666a767542121f5683131117b89c74e",
                "parameters": [
                    {
                        "name": "answer",
                        "in": "path",
                        "description": "The ID of the answer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Answer deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the answer was deleted successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Answer deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Answer": {
                "title": "Answer",
                "description": "Answer model schema",
                "required": [
                    "answer",
                    "user_id",
                    "question_id"
                ],
                "properties": {
                    "id": {
                        "description": "The ID of the answer",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "answer": {
                        "description": "The content of the answer",
                        "type": "string",
                        "example": "This is a detailed answer to the question."
                    },
                    "user_id": {
                        "description": "The ID of the user who wrote the answer",
                        "type": "integer",
                        "example": 5
                    },
                    "question_id": {
                        "description": "The ID of the question being answered",
                        "type": "integer",
                        "example": 12
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-11T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-11T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}